+++ src/test/java/racinggame/domain/StepTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racinggame.utils.Message;
+
 class StepTest {
 
     @DisplayName("스텝 메시지 생성 테스트")
