+++ src/test/java/racinggame/utils/MessageBuilderTest.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.utils;
 
 import static org.assertj.core.api.Assertions.*;
